chore: rewrite FrankerFaceZ
