+++ layers/emote-integrations/layers/integrations/frankerfacez/server/utils/types.global.ts
@@ -1,0 +1,12 @@
+export type Emote = TFrankerFaceZ.GlobalEmote;
+
+export type Set = TFrankerFaceZ.GlobalEmoteSet;
+
+export interface ReadyIntegration extends TEmoteIntegrations.Global.Ready {
+  sets: GlobalEmoteSet[];
+  source: "FrankerFaceZ";
+}
+
+export interface FailedIntegration extends TEmoteIntegrations.Global.Failed {
+  source: "FrankerFaceZ";
+}
